+++ package.json
@@ -38,7 +38,10 @@
     "catalog-volume-monitoring:build": "yarn workspace @akeneo-pim-community/catalog-volume-monitoring lib:build",
     "config:lint:check": "yarn workspace @akeneo-pim-community/config lint:check",
     "config:build": "yarn workspace @akeneo-pim-community/config lib:build",
-    "packages:unit": "yarn workspace @akeneo-pim-community/shared test:unit:run && yarn workspace @akeneo-pim-community/measurement test:unit:run && yarn workspace @akeneo-pim-community/process-tracker test:unit:run && yarn workspace @akeneo-pim-community/catalog-volume-monitoring test:unit:run",
+    "category:lint:check": "yarn workspace @akeneo-pim-community/category lint:check",
+    "category:build": "yarn workspace @akeneo-pim-community/category lib:build",
+    "category:unit": "yarn workspace @akeneo-pim-community/category test:unit:run",
+    "packages:unit": "yarn workspace @akeneo-pim-community/shared test:unit:run && yarn workspace @akeneo-pim-community/measurement test:unit:run && yarn workspace @akeneo-pim-community/process-tracker test:unit:run && yarn workspace @akeneo-pim-community/catalog-volume-monitoring test:unit:run && yarn workspace @akeneo-pim-community/category test:unit:run",
     "packages:lint:check": "yarn shared:lint:check && yarn measurement:lint:check && yarn process-tracker:lint:check && yarn catalog-volume-monitoring:lint:check && yarn config:lint:check",
     "packages:build": "yarn dsm:build && yarn shared:build && yarn measurement:build && yarn catalog-volume-monitoring:build && yarn config:build && yarn process-tracker:build  && yarn catalogs:build",
     "test:e2e:open": "cypress open",
@@ -178,6 +181,7 @@
     "src/Akeneo/Pim/Enrichment/Bundle/Resources/workspaces/enrichment",
     "src/Akeneo/Platform/Bundle/CatalogVolumeMonitoringBundle/front",
     "src/Oro/Bundle/ConfigBundle/front",
+    "src/Akeneo/Category/front",
     "components/catalogs/front"
   ]
 }
