+++ src/Akeneo/Category/front/.gitignore
@@ -1,4 +1,5 @@
-/node_modules
+node_modules
+coverage
 routes.json
 translations.json
 .eslintcache
