+++ src/Akeneo/Category/front/package.json
@@ -44,5 +44,6 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
-    }
+    },
+    "jest": {}
 }
