+++ src/Akeneo/Category/front/src/feature/helpers/permissionsHelper.test.ts
@@ -1,9 +1,6 @@
-import {
-  computeNewEditPermissions,
-  computeNewOwnPermissions,
-  computeNewViewPermissions,
-} from '@akeneo-pim-community/settings-ui/src/helpers';
-import {EditCategoryForm} from '@akeneo-pim-community/settings-ui/src';
+import { EditCategoryForm } from "feature/models";
+import { computeNewEditPermissions, computeNewOwnPermissions, computeNewViewPermissions } from "./permissionsHelper";
+
 
 describe('permissionsHelper', () => {
   //This test is only to have a better idea of the default state before each test
@@ -64,7 +61,7 @@ const formData: EditCategoryForm = {
   },
   errors: [],
   _token: {
-    label: 'token',
+    //label: 'token',
     value: 'XFC_DnwJvzF5TsnB2-MbiPUjKqmUtZTJt0O1CTQLqMs',
     fullName: 'pim_category[_token]',
   },
@@ -85,7 +82,7 @@ const formData: EditCategoryForm = {
       choices: [],
     },
     apply_on_children: {
-      label: '',
+      //label: '',
       value: '1',
       fullName: 'pim_category[permissions][apply_on_children]',
     },
