+++ src/Akeneo/Category/front/src/feature/hooks/useDragTreeNode.test.tsx
@@ -1,19 +1,28 @@
 import React, {FC} from 'react';
-import {TreeNode, useDragTreeNode} from '../../../../../src';
-import {DependenciesProvider} from '@akeneo-pim-community/legacy-bridge';
 import {ThemeProvider} from 'styled-components';
 import {pimTheme} from 'akeneo-design-system';
 import {renderHook} from '@testing-library/react-hooks';
-import {OrderableTreeProvider} from '../../../../../src/components/shared/providers/OrderableTreeProvider';
-import {aTreeNode} from '../../../utils/provideTreeNodeHelper';
 import {act} from 'react-test-renderer';
+import { OrderableTreeProvider } from 'feature/components';
+import { TreeNode } from 'feature/models';
+import { useDragTreeNode } from './useDragTreeNode';
+import { aTreeNode } from 'tests/provideTreeNodeHelper';
+import { MicroFrontendDependenciesProvider, Routes, Translations } from '@akeneo-pim-community/shared';
+// import {routes} from '../../routes.json';
+// import translations from '../../translations.json';
+
+const routes: Routes = {
+  pim_user_user_rest_get_current: {tokens: []},
+  pim_user_security_rest_get: {tokens: []},
+}
 
 const DefaultProviders: FC<{orderable: boolean}> = ({children, orderable}) => (
-  <DependenciesProvider>
-    <ThemeProvider theme={pimTheme}>
+//  <MicroFrontendDependenciesProvider  routes={routes as Routes} translations={translations as Translations}>
+<MicroFrontendDependenciesProvider  routes={routes} translations={{locale:'en-US', messages: {}}}>
+<ThemeProvider theme={pimTheme}>
       <OrderableTreeProvider isActive={orderable}>{children}</OrderableTreeProvider>
     </ThemeProvider>
-  </DependenciesProvider>
+  </MicroFrontendDependenciesProvider>
 );
 
 const renderUseDragTreeNode = (node: TreeNode<any> | undefined, index: number, orderable: boolean) => {
