+++ src/Akeneo/Category/front/tsconfig.json
@@ -18,7 +18,7 @@
     "noUnusedLocals": false,
     "noUnusedParameters": false,
     "outDir": "lib",
-    "rootDir": "src/feature",
+    "rootDir": "src",
     "skipLibCheck": true,
     "sourceMap": true,
     "strictBindCallApply": true,
