--- src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './treeHelper';
