+++ src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/hooks/index.ts
@@ -2,4 +2,3 @@ export * from './shared';
 export * from './attribute-groups';
 export * from './locales';
 export * from './association-types';
-export * from './categories';
